+++ mode/guozhan/src/character/index.js
@@ -5,6 +5,7 @@ import zhen from "./zhen.js";
 import shi from "./shi.js";
 import bian from "./bian.js";
 import quan from "./quan.js";
+import yingbian from "./yingbian.js";
 
 import restCharacter from "./rest.js";
 
@@ -143,6 +144,7 @@ export const pack = {
 	...shi,
 	...bian,
 	...quan,
+	...yingbian,
 };
 
 export { default as intro } from "./intro.js";
