+++ mode/guozhan/src/character/rest.js
@@ -1,7 +1,4 @@
 import { Character } from "../../../../noname/library/element/index.js";
-import { _status, get } from "../../../../noname.js";
-
-const bool = _status.connectMode ? lib.configOL.jinEx : get.config("jinEx");
 
 export default {
 	// 分组1：野心家
@@ -523,247 +520,6 @@ export default {
 		hujia: 0,
 		skills: ["jianan", "huibian", "gzzongyu"],
 	}),
-	// 分组3：晋势力
-	gz_jun_jin_simayi: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_jiaping", "gz_guikuang", "gz_shujuan"],
-	}),
-	gz_simaliang: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_gongzhi", "gz_sheju"],
-		img: "image/character/jsrg_simaliang.jpg",
-	}),
-	gz_wangjun: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_chengliu"],
-		img: "image/character/jsrg_wangjun.jpg",
-	}),
-	gz_malong: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_zhuanzhan", "gz_xunji"],
-		img: "image/character/jsrg_malong.jpg",
-	}),
-	gz_simalun: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_zhulan", "gz_luanchang"],
-	}),
-	gz_jin_guohuai: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_zhefu", "gz_yidu"],
-	}),
-	gz_wenyang: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 5,
-		maxHp: 5,
-		hujia: 0,
-		skills: ["gz_duanqiu"],
-		hasSkinInGuozhan: true,
-	}),
-	gz_bailingyun: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_xiace", "gz_limeng"],
-		hasSkinInGuozhan: true,
-		names: "柏|灵筠",
-	}),
-	gz_sunxiù: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_xiejian", "gz_yinsha"],
-	}),
-	gz_yangjun: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_neiji"],
-	}),
-	gz_wangxiang: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_bingxin"],
-	}),
-	gz_jin_zhangchunhua: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["gzhuishi", "fakeqingleng"] : ["gz_ejue", "gz_shangshi"],
-		hasSkinInGuozhan: bool ? null : true,
-	}),
-	gz_jin_simayi: new Character({
-		sex: "male",
-		group: "jin",
-		hp: bool ? 3 : 4,
-		maxHp: bool ? 3 : 4,
-		hujia: 0,
-		skills: bool ? ["fakequanbian", "smyyingshi", "fakezhouting"] : ["gz_yingshi", "gz_shunfu"],
-		hasSkinInGuozhan: bool ? null : true,
-	}),
-	gz_jin_wangyuanji: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["fakeyanxi", "fakeshiren"] : ["gz_yanxi", "gz_shiren"],
-	}),
-	gz_jin_simazhao: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["zhaoran", "gzchoufa"] : ["gz_zhaoran", "gz_beiluan"],
-	}),
-	gz_jin_xiahouhui: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["fakebaoqie", "jyishi", "shiduo"],
-	}),
-	gz_jin_simashi: new Character({
-		sex: "male",
-		group: "jin",
-		hp: bool ? 5 : 4,
-		maxHp: bool ? 5 : 4,
-		hujia: 0,
-		skills: bool ? ["gzyimie", "gztairan"] : ["gz_yimie", "gz_ruilve"],
-	}),
-	gz_duyu: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: bool ? ["gzsanchen", "gzpozhu"] : ["gz_sanchen", "gz_pozhu"],
-	}),
-	gz_zhanghuyuechen: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: bool ? ["fakexijue"] : ["gz_xijue", "gz_lvxian", "gz_yingwei"],
-	}),
-	gz_jin_yanghuiyu: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["fakeciwei", "fakehuirong"] : ["gz_ciwei", "gz_caiyuan"],
-	}),
-	gz_simazhou: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: bool ? ["caiwang", "gznaxiang"] : ["gz_pojing"],
-	}),
-	gz_shibao: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: bool ? ["gzzhuosheng"] : ["gz_zhuosheng"],
-	}),
-	gz_weiguan: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["zhongyun", "shenpin"] : ["gz_chengxi", "gz_jiantong"],
-	}),
-	gz_zhongyan: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gzbolan", "yifa"],
-	}),
-	gz_yangyan: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gzxuanbei", "xianwan"],
-	}),
-	gz_zuofen: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gzzhaosong", "gzlisi"],
-	}),
-	gz_xuangongzhu: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["fakeqimei", "ybzhuiji"],
-	}),
-	gz_xinchang: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["fakecanmou", "congjian"],
-	}),
-	gz_yangzhi: new Character({
-		sex: "female",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gzwanyi", "gzmaihuo"],
-	}),
 
 	gz_liaohua: new Character({
 		sex: "male",
@@ -954,22 +710,6 @@ export default {
 		hujia: 0,
 		skills: ["fakexunxi", "fakehuanjia"],
 		hasSkinInGuozhan: true,
-	}),
-	gz_jin_jiachong: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: bool ? ["fakexiongshu", "fakejianhui"] : ["gz_chujue", "gz_jianzhi"],
-	}),
-	gz_jin_yanghu: new Character({
-		sex: "male",
-		group: "jin",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: bool ? ["fakechongxin", "fakeweirong"] : ["gz_huaiyuan", "gz_fushou"],
 	}),
 };
