+++ mode/guozhan/src/skill/character/index.js
@@ -3,6 +3,7 @@ import zhen from "./zhen.js";
 import shi from "./shi.js";
 import bian from "./bian.js";
 import quan from "./quan.js";
+import yingbian from "./yingbian.js";
 
 import rest from "./rest.js";
 
@@ -13,4 +14,5 @@ export default {
 	...shi,
 	...bian,
 	...quan,
+	...yingbian,
 };
