+++ mode/guozhan/src/translate/rest.js
@@ -4,16 +4,12 @@ export default {
 	gzqiangxi: "强袭",
 	gzqiangxi_info: "出牌阶段限一次，你可以弃置一张武器牌或失去1点体力，然后对一名其他角色造成1点伤害。",
 
-	gz_sp_duyu: "杜预",
 	gzpozhen: "破阵",
 	gzpozhen_info: "限定技，其他角色的回合开始时，你可以令其本回合不可使用、打出或重铸手牌；若其处于队列或围攻关系中，你可依次弃置此队列或参与围攻关系的其他角色的一张牌。",
 	gzjiancai: "荐才",
 	gzjiancai_info: "副将技，此武将牌上单独的阴阳鱼个数-1。与你势力相同的角色即将受到伤害而进入濒死状态时，你可以防止此伤害，若如此做，你须变更副将；与你势力相同的角色变更副将时，其额外获得两张备选武将牌。",
 	gzzhuhai: "诛害",
 	gzzhuhai_info: "其他角色的结束阶段，若其本回合内造成过伤害，则你可以对其使用一张【杀】（无距离限制）。若其本回合内对与你势力相同的角色造成过伤害，则此【杀】无视防具，且当其抵消此【杀】后，其须弃置一张牌。",
-	gzzhuosheng: "擢升",
-	gzzhuosheng2: "擢升",
-	gzzhuosheng_info: "当一名与你势力相同的角色受到伤害后，你可令其摸一张牌。然后直到其下个结束阶段前，其使用此牌根据类型执行以下效果：1. 基本牌，不计入次数且无距离限制；2. 普通锦囊牌，此牌目标可+1或-1；3. 装备牌，可摸一张牌。",
 
 	gzzhaoxin: "昭心",
 	gzzhaoxin_info: "当你受到伤害后，你可展示所有手牌，然后与一名手牌数不大于你的其他角色交换手牌。",
@@ -358,37 +354,6 @@ export default {
 	gzhongyuan_info: "①出牌阶段限一次。你可以令一张没有「合纵」标签的卡牌视为拥有「合纵」标签直到本回合结束。②当你即将因合纵效果摸牌时，你可放弃摸牌，并令一名己方角色摸等量的牌。",
 	gzmingzhe: "明哲",
 	gzmingzhe_info: "你的回合外，当你使用或打出红色手牌，或失去装备区内的红色装备牌时，你可摸一张牌。",
-	gzquanbian: "权变",
-	gzquanbian_info: "当你于出牌阶段内使用/打出手牌时，若此牌有花色且你本回合内未使用/打出过该花色的其他手牌，则你可以观看牌堆顶X张牌，选择获得其中的一张并展示之。若你本回合使用过与得到的牌花色相同的牌，则你本回合内不能再发动〖权变〗。",
-	gzhuishi: "慧识",
-	gzhuishi_info: "摸牌阶段，你可以放弃摸牌，改为观看牌堆顶的X张牌，获得其中的一半（向下取整），然后将其余牌置入牌堆底。（X为弃牌堆顶的两张牌的名称字数之和）",
-	gzqingleng: "清冷",
-	gzqingleng_info: "一名角色的回合结束时，若其有未明置的武将牌，则你可将一张牌当无距离限制的冰属性【杀】对其使用；若其所有武将牌均未明置，则你摸一张牌。",
-	gzchoufa: "筹伐",
-	gzchoufa_info: "出牌阶段限一次，你可展示一名其他角色的一张手牌A。你令其所有类型与A不同的手牌的牌名均视为【杀】且均视为无属性直到本回合结束。",
-	gzbolan: "博览",
-	gzbolan_info: "每名角色的出牌阶段限一次。其可以随机展示武将牌堆顶的一张武将牌，然后根据此武将牌包含的势力选择获得一个技能直到回合结束：魏:〖奇策〗；蜀:〖挑衅〗；吴:〖制衡〗；群:〖除疠〗；晋:〖三陈〗。若该角色不为你，则其失去1点体力。",
-	gzsanchen: "三陈",
-	gzsanchen_info: "出牌阶段，你可选择一名本回合内未选择过的角色。其摸三张牌，然后弃置三张牌。若其未以此法弃置牌或以此法弃置的牌的类别均不相同，则你于本回合获得一枚“陈”且其摸一张牌；否则你本阶段内不能再发动〖三陈〗。",
-	gzpozhu: "破竹",
-	gzpozhu_info: "主将技。此武将牌减少半个阴阳鱼。出牌阶段，你可以移去一枚“陈”并将一张手牌当做【出其不意】使用。",
-	gzyimie: "夷灭",
-	gzyimie_info: "主将技，此武将牌减少一整个阴阳鱼。每回合限一次，当你对其他角色造成伤害时，若伤害值X小于Y，则你可失去1点体力，将伤害值改为Y。此伤害结算结束后，其回复(Y-X)点体力。（Y为其体力值）",
-	gztairan: "泰然",
-	gztairan_info: "出牌阶段开始时，你可以失去任意点体力并弃置任意张牌（弃牌数不能大于你已损失的体力值）。若如此做，本回合结束时，你回复等量的体力并摸等量的牌。",
-	gzxuanbei: "选备",
-	gzxuanbei_info: "①当你首次明置此武将牌时，你获得两张拥有“应变”或“合纵”标签的牌（不足则摸牌），且你本回合内使用带有“应变”标签的牌时直接强化，使用带有“合纵”标签的牌时摸一张牌。②当你死亡时，你可以令一名其他角色变更副将。",
-	gzzhaosong: "诏颂",
-	gzzhaosong_info: "每局游戏每项限一次。①一名角色进入濒死状态时，你可以令其回复至2点体力并摸一张牌。②出牌阶段，你可观看一名其他角色的所有暗置武将牌和手牌，然后可以获得其区域内的一张牌。③一名角色使用【杀】选择唯一目标后，你可以为此【杀】增加两个目标。",
-	gzlisi: "离思",
-	gzlisi_info: "一名己方角色死亡后，你可以选择〖诏颂〗中的一个已发动过的选项，令其视为未发动过。",
-	gzcaiyuan: "才媛",
-	gzcaiyuan_info: "锁定技。结束阶段，若你的手牌数大于本回合开始时的手牌数，则你摸两张牌或回复1点体力。",
-	gzwanyi: "婉嫕",
-	gzwanyi_info: "出牌阶段每项各限一次。你可以将一张带有“合纵”标签的牌当做【联军盛宴】/【火烧连营】/【挟天子以令诸侯】/【戮力同心】使用。",
-	gzwanyi_info_guozhan_yingbian: "出牌阶段每项各限一次。你可以将一张带有“应变”标签的牌当做【逐近弃远】/【洞烛先机】/【水淹七军】/【出其不意】使用。",
-	gzmaihuo: "埋祸",
-	gzmaihuo_info: "限定技。当有己方角色成为【杀】的目标时，你可以取消此【杀】的所有目标。然后此【杀】的使用者下回合开始时，其视为对你使用一张【杀】。若此【杀】对你造成伤害，则你防止此伤害，摸两张牌并移除此武将牌（若此武将牌为副将则改为变更副将）。",
 	gzzhenxi: "震袭",
 	gzzhenxi_info: "当你使用【杀】指定目标后，你可以选择一项：1.弃置目标角色一张牌；2.将一张♦非锦囊牌当做【乐不思蜀】或♣非锦囊牌当做【兵粮寸断】对目标角色使用；3.背水：若其有暗置的武将牌且你的武将牌均明置，则你可以依次执行这两项。",
 	gzjiansu: "俭素",
@@ -541,56 +506,9 @@ export default {
 	fakexunxi_info: "其他角色于回合外明置武将牌时，你可以视为对其使用一张【杀】（无距离限制）。",
 	fakehuanjia: "擐甲",
 	fakehuanjia_info: "锁定技，每回合每项各限一次。①当你成为【杀】的目标后，本回合你视为装备此牌使用者的防具，直到你的装备区中有防具。②当你使用【杀】指定唯一目标后，本回合你视为装备目标角色的武器，直到你的装备区中有武器。",
-	fakexiongshu: "凶竖",
-	fakexiongshu_info: "每回合限一次，一名角色使用【杀】或伤害类锦囊牌指定第一个目标后，若你为/不为此牌使用者，则你可以摸一张牌并令一名其他角色成为此牌的伤害来源/弃置一张牌并成为此牌的伤害来源。",
-	fakejianhui: "奸回",
-	fakejianhui_info: "当一名角色受到其势力相同的另一名角色造成的伤害后，你可以令其中一名角色摸一张牌，然后弃置其中另一名角色的一张牌。",
-	fakechongxin: "崇信",
-	fakechongxin_info: "出牌阶段限一次，你可以选择一名与你势力不同的角色，视为你使用一张指定你与其为目标的【以逸待劳】。",
-	fakeweirong: "卫戎",
-	fakeweirong_info: "转换技，每轮限一次，出牌阶段。阴：你可以弃置X张牌，然后当你于本轮不因此法得到牌后，你摸一张牌。阳：你可以摸X张牌，然后当你于本轮不因此法失去牌后，你弃置一张牌。（X为你上一轮以此法摸和弃置的牌数之和，且X至少为1，至多为你的体力上限）",
 	gz_re_taishici: "太史慈",
 	gz_xin_zhuran: "朱然",
-	gz_jin_jiachong: "贾充",
-	gz_jin_yanghu: "羊祜",
 	gz_wangling: "王凌",
-	fakequanbian: "权变",
-	fakequanbian_tag: "牌堆顶",
-	fakequanbian_info: "当你于回合内使用或打出牌时，你可以将一张手牌与牌堆顶X张牌的其中一张进行交换（X为你的体力上限）。",
-	fakezhouting: "骤霆",
-	fakezhouting_info: "限定技，出牌阶段，你可以依次使用牌堆顶X张牌中所有可以使用的牌，然后获得其中不能使用的牌，然后若有角色因此死亡，则你重置〖骤霆〗（X为你的体力上限）。",
-	fakexuanbei: "选备",
-	fakexuanbei_info: "①当你首次明置此武将牌时，你摸两张牌，且你本回合内使用带有“应变”标签的牌时直接强化。②每回合限一次，当你使用带有“应变”标签的牌结算完毕后，你可以令一名其他角色获得此牌对应的所有实体牌。③当你死亡时，你可以令一名其他角色变更副将。",
-	fakeqingleng: "清冷",
-	fakeqingleng_info: "其他角色的结束阶段，若其武将牌均明置，则你可以将一张牌当作冰【杀】对其使用（无距离限制）。然后若此牌未对其造成伤害，则你暗置其一张武将牌，且直到其下个回合开始，其明置此武将牌时，你对其造成1点伤害。",
-	fakexijue: "袭爵",
-	fakexijue_info: "①当你首次明置此武将牌时，你获得2枚“爵”标记。②回合结束时，若你本回合造成过伤害，则你获得1枚“爵”标记。③你可以于合适的时机发动〖突袭〗或〖骁果〗（至多弃置X张基本牌），然后移去1枚“爵”标记。④摸牌阶段，你改为摸X张牌。（X为你的“爵”标记数，且X至多为你的体力上限）",
-	fakexijue_xiaoguo: "骁果",
-	fakexijue_xiaoguo_info: "一名其他角色的准备阶段，你可以弃置至多X张基本牌（X为你的“爵”标记数，且X至多为你的体力上限），然后弃置其装备区等量的牌，若其装备区的牌数小于你弃置的牌数，则你对其造成1点伤害。",
-	fakeqimei: "齐眉",
-	fakeqimei_info: "准备阶段，你可以选择一名其他角色。若如此做，直到回合结束：当你或其获得牌/失去手牌后，若你与其手牌数相等，则另一名角色回复1点体力；当你或其的体力值变化后，若你与其体力值相等，则另一名角色摸一张牌。",
-	fakebaoqie: "宝箧",
-	fakebaoqie_info: "①当你受到伤害时，若此武将牌未明置过，则你可以明置此武将牌并防止此伤害。②当你首次明置此武将牌时，你可以获得一名角色装备区里所有的宝物牌，然后你可以使用其中的一张牌。",
-	fakeciwei: "慈威",
-	fakeciwei_info: "其他角色于一回合内使用第偶数张牌时，你可以弃置一张牌并取消此牌的所有目标，然后若此牌为装备牌，你可以获得之且你可以使用之。",
-	fakehuirong: "慧容",
-	fakehuirong_info: "①当你受到伤害时，若此武将牌未明置过，则你可以明置此武将牌并防止此伤害。②当你首次明置此武将牌时，你可以令一名角色将手牌数调整至其体力值。",
-	fakeyanxi: "宴戏",
-	fakeyanxi_info: "出牌阶段限一次，你可以展示一名其他角色的一张手牌和牌堆顶的两张牌，然后你获得其中一种颜色或类别的所有牌且本回合手牌上限+X（X为你获得的牌数），然后其获得剩余的牌。",
-	fakeshiren: "识人",
-	fakeshiren_info: "①当你受到伤害时，若此武将牌未明置过，则你可以明置此武将牌并防止此伤害。②当你首次明置此武将牌时，你可以发动一次〖宴戏〗。",
-	fakecanmou: "参谋",
-	fakecanmou_info: "每回合每项限一次，一名角色使用普通锦囊牌指定第一个目标时，若其手牌数为全场唯一最多，则你可以为此牌增加或减少一个目标（目标数至少为1）。",
-	fakezhuosheng: "擢升",
-	fakezhuosheng_info: "每回合每种牌名限一次，你可以将至少两张手牌当作任意基本牌使用，若你以此法使用的牌中包含具有应变效果的牌，则你令此牌无视条件获得对应的应变效果。",
-	fakejuhou: "拒后",
-	fakejuhou_info: "阵法技，与你处于同一队列的角色成为【杀】或普通锦囊牌的目标后，你可以令其将任意张牌置于其武将牌上，然后其于此牌结算完毕后获得这些牌。",
-	gznaxiang: "纳降",
-	gznaxiang_info: "锁定技，当你受到其他角色造成的伤害后，或你对其他角色造成伤害后，你对其发动〖才望〗时的“弃置”改为“获得”直到你的下回合开始。",
-	fakecaiwang: "才望",
-	fakecaiwang_info: "①你可以助战自己。②每回合每项各限一次，当你响应助战后，若你弃置的牌和被强化的牌：颜色相同，你可以弃置一名其他角色的一张牌；颜色不同，你摸一张牌。",
-	fakenaxiang: "纳降",
-	fakenaxiang_info: "锁定技。①当你对与你势力不同的角色造成伤害后，或受到与你势力不同的角色对你造成的伤害后。你对其发起军令，若其不执行，则你不能对其发动〖纳降①〗直到你的下个回合开始。②你响应助战弃牌无类别限制，且你触发具有应变的卡牌的条件均视为助战。",
 
 	fakekeshou: "恪守",
 	fakekeshou_info: "①当你受到伤害时，你可以弃置两张颜色相同的牌并令此伤害-1。②当你因弃置而一次性失去至少两张牌后，若你的势力已确定且场上没有与你势力相同的其他角色，则你可以进行判定，若结果判定为红色，你摸一张牌。",
@@ -628,116 +546,6 @@ export default {
 
 	gzchengshang: "承赏",
 	gzchengshang_info: "每回合限一次，当你使用指定了目标的牌结算完毕后，若你未因此牌造成过伤害，则你可以令其中一名目标角色交给你一张牌，若此牌和你使用的牌的花色和点数均相同，则你失去此技能。",
-
-	gz_jun_jin_simayi: "君司马懿",
-	gz_jun_jin_simayi_prefix: "君",
-	gz_jiaping: "嘉平",
-	gz_jiaping_info: "君主技，只要此武将牌处于明置状态，你便拥有“八荒死士令”。",
-	bahuangsishiling: "八荒死士令",
-	bahuangsishiling_ab: "士令",
-	bahuangsishiling_bg: "令",
-	bahuangsishiling_info: "每轮共计限一次，本轮明置过武将牌的晋势力角色可以于对应时机移除副将并发动一个未以此法发动过的技能：〖瞬覆〗〖乱武〗〖将略〗〖勇进〗〖奉迎〗。",
-	gz_guikuang: "诡诳",
-	gz_guikuang_info: "出牌阶段限一次，你可以令两名势力不相同的角色拼点；拼点牌为红色的角色依次对没赢的角色造成1点伤害。",
-	gz_shujuan: "舒卷",
-	gz_shujuan_info: "锁定技，【戢鳞潜翼】每回合首次进入弃牌堆或其他角色装备区时，你获得并使用之。",
-	gz_wenyang: "文鸯",
-	gz_duanqiu: "断虬",
-	gz_duanqiu_info: "准备阶段，你可以视为对一个势力的所有其他角色使用一张【决斗】。结算后令所有角色本回合仅能合计再使用X张手牌（X为此【决斗】结算过程中打出【杀】的数量）。",
-	gz_bailingyun: "柏夫人",
-	gz_xiace: "黠策",
-	gz_xiace_info: "若当前回合角色有出牌阶段剩余出杀次数，你可以将一张牌当【无懈可击】使用并令其本回合出牌阶段出杀次数-1，然后你可以变更副将。",
-	gz_limeng: "离梦",
-	gz_limeng_info: "结束阶段，你可以弃置一张非基本牌并选择两张珠联璧合的武将牌，若不为一名角色的武将牌，这些角色依次对对方造成1点伤害。",
-	gz_sunxiù: "孙秀",
-	gz_xiejian: "挟奸",
-	gz_xiejian_info: "出牌阶段限一次，你可以对一名其他角色发起一次“军令”，若其不执行，其执行未被你选择的另一项“军令”。",
-	gz_yinsha: "引杀",
-	gz_yinsha_info: "你可以将所有手牌当【借刀杀人】使用，目标角色须使用【杀】响应（无【杀】则改为将所有手牌当【杀】使用）。",
-	gz_yangjun: "杨骏",
-	gz_neiji: "内忌",
-	gz_neiji_info: "出牌阶段开始时，你可以选择一名与你势力不同或未确定势力的其他角色，与其同时展示两张手牌，然后弃置展示的【杀】。若合计弃置牌数大于1，你与其各摸三张牌；否则未弃置牌的角色视为对弃置了牌的角色使用一张【决斗】。",
-	gz_wangxiang: "王祥",
-	gz_bingxin: "冰心",
-	gz_bingxin_info: "每种牌名每回合限一次。当你需要使用基本牌时，若你的手牌数等于体力值且这些牌的颜色均相同，则你可以摸一张牌，视为使用一张基本牌。",
-	gz_sanchen: "三陈",
-	gz_sanchen_info: "出牌阶段每名角色限一次，若你的武将牌均明置，你可以令一名角色摸三张牌并弃置三张牌，若其因此弃置了相同类型的牌，你暗置此武将牌。",
-	gz_pozhu: "破竹",
-	gz_pozhu_info: "主将技。①此武将牌计算体力上限时减少半个阴阳鱼。②准备阶段，你可以将一张牌当【杀】使用，结算后展示此牌唯一目标的一张手牌，若两张牌花色不同，你可以重复此流程。",
-	gz_huaiyuan: "怀远",
-	gz_huaiyuan_info: "与你势力相同角色的准备阶段开始时，你可以令其本回合一项数值+1：攻击范围、手牌上限、出【杀】次数。",
-	gz_fushou: "付授",
-	gz_fushou_info: "锁定技，与你势力相同的角色无视主副将条件拥有其所有主将技和副将技（不重新计算阴阳鱼）。",
-	gz_xijue: "袭爵",
-	gz_xijue_info: "你可以弃置一张牌并发动〖突袭〗或〖骁果〗。",
-	gz_xijue_tuxi: "突袭",
-	gz_xijue_tuxi_info: "摸牌阶段，你可以少摸任意张牌，获得等量其他角色各一张手牌。",
-	gz_xijue_xiaoguo: "骁果",
-	gz_xijue_xiaoguo_info: "其他角色的结束阶段，你可以弃置一张基本牌，令该角色选择一项：1.弃置一张装备牌；2.受到你对其造成的1点伤害。",
-	gz_lvxian: "履险",
-	gz_lvxian_info: "主将技，你每回合首次受到伤害后，若执行上回合的角色不为你，你可以摸你上回合失去牌数张牌。",
-	gz_yingwei: "盈威",
-	gz_yingwei_info: "副将技，结束阶段，若你本回合摸牌数等于造成伤害数，你可以重铸至多两张牌。",
-
-	gz_yingshi: "鹰视",
-	gz_yingshi_info: "出牌阶段开始时，你可以令一名角色视为对你指定的令一名角色使用一张【知己知彼】，然后若使用者不为你，你摸一张牌。",
-	gz_shunfu: "瞬覆",
-	gz_shunfu_info: "限定技，出牌阶段，你可以令至多三名未确定势力的其他角色各摸两张牌，这些角色依次可以使用一张无距离限制且不可被响应的【杀】。",
-	gz_ejue: "扼绝",
-	gz_ejue_info: "锁定技，你使用【杀】对一名角色造成伤害时，若其未确定势力，此伤害+1。",
-	gz_shangshi: "伤逝",
-	gz_shangshi_info: "一名角色的回合结束时，你可以将手牌摸至你的已损失体力值",
-	gz_yimie: "夷灭",
-	gz_yimie_info: "锁定技，你回合内的濒死结算中，与濒死角色势力相同的角色不能使用桃，不同的角色可以将一张♥手牌当【桃】使用。",
-	gz_ruilve: "睿略",
-	gz_ruilve_info: "未确定势力的其他角色出牌阶段限一次，其可以展示并交给你一张伤害牌，然后摸一张牌。",
-	gz_zhaoran: "昭然",
-	gz_zhaoran_info: "出牌阶段开始前，你可以摸4-X张牌（X为场上势力数），然后未确定势力的角色可以明置一张武将牌并令你结束此回合。",
-	gz_beiluan: "备乱",
-	gz_beiluan_info: "你受到伤害后，你可以令伤害来源所有非装备手牌均视为【杀】直到此回合结束。",
-	gz_pojing: "迫境",
-	gz_pojing_info: "出牌阶段限一次，你可以令一名其他角色选择一项：1.令你获得其区域里的一张牌；2.所有与你势力相同的角色可以明置任意张武将牌并对其造成等量伤害。",
-	gz_simaliang: "司马亮",
-	gz_gongzhi: "共执",
-	gz_gongzhi_info: "你可以跳过摸牌阶段，令与你势力相同的角色依次摸一张牌直到共计摸四张牌。",
-	gz_sheju: "慑惧",
-	gz_sheju_info: "锁定技，其他角色明置武将牌后，若与你势力相同，你回复1点体力，然后弃置所有手牌。",
-	gz_simalun: "司马伦",
-	gz_zhulan: "助澜",
-	gz_zhulan_info: "一名其他角色受到伤害时，若伤害来源与其同势力，你可以弃置一张牌令此伤害+1。",
-	gz_luanchang: "乱常",
-	gz_luanchang_info: "限定技，一名角色的回合结束时，若有本回合受到伤害的角色与你势力相同，你可以令当前回合角色将所有手牌当【万箭齐发】使用。",
-	gz_zhuosheng: "擢升",
-	gz_zhuosheng_info: "当你获得牌后，你可以令你本回合使用的下一张牌造成的伤害+1（不可叠加）。",
-	gz_ciwei: "慈威",
-	gz_ciwei_info: "你的回合内，其他角色使用牌时，若有和其势力不同的其他角色本回合使用或打出过牌，你可以弃置一张牌令其无效。",
-	gz_caiyuan: "才媛",
-	gz_caiyuan_info: "锁定技，若你的武将牌均明置：回合开始时，你摸两张牌；你受到伤害后，暗置此武将牌。",
-	gz_yanxi: "宴戏",
-	gz_yanxi_info: "准备阶段，你可以选择至多三名与你势力不相同角色的各一张牌，这些角色依次声明一个牌名，然后你展示并获得其中一张，若此牌牌名与其声明的不同，你获得未被选择的牌。",
-	gz_shiren: "识人",
-	gz_shiren_info: "每回合限一次，未确定势力的其他角色受到伤害后，你可以交给其两张牌并摸两张牌。",
-	gz_chengxi: "乘隙",
-	gz_chengxi_info: "准备阶段，你可令一名角色视为使用一张【以逸待劳】，结算完成后若以此法弃置的牌包含非基本牌，此牌使用者对所有目标角色各造成1点伤害。",
-	gz_jiantong: "监统",
-	gz_jiantong_info: "你受到伤害后，可以观看一名角色的手牌，然后可以用装备区里的一张牌交换其中至多两张牌。",
-	gz_chujue: "除绝",
-	gz_chujue_info: "锁定技，你对有角色死亡的势力使用牌无次数限制且不可被该势力的角色响应。",
-	gz_jianzhi: "奸志",
-	gz_jianzhi_info: "你造成伤害时，若伤害值不小于受伤角色体力值，你可以弃置所有手牌，令本回合下次击杀奖励改为三倍。",
-	gz_jin_guohuai: "郭槐",
-	gz_zhefu: "哲妇",
-	gz_zhefu_info: "你于回合外使用或打出基本牌时，可以选择一名与其同势力角色数不小于与你同势力角色数的角色，观看其手牌，然后可以弃置其中一张基本牌。",
-	gz_yidu: "遗毒",
-	gz_yidu_info: "你使用伤害类牌后，可以展示一名未受到此牌伤害的目标角色至多两张手牌，若颜色相同，你弃置这些牌。",
-	gz_wangjun: "王濬",
-	gz_chengliu: "乘流",
-	gz_chengliu_info: "出牌阶段限一次，你可以对一名装备区牌数小于你的角色造成1点伤害，然后你可以与其交换装备区内的牌并重复此流程。",
-	gz_malong: "马隆",
-	gz_zhuanzhan: "转战",
-	gz_zhuanzhan_info: "锁定技，若场上有未确定势力的角色，你使用【杀】无距离限制且不能指定未确定势力的角色为目标。",
-	gz_xunji: "勋济",
-	gz_xunji_info: "你使用【杀】可以额外指定两名目标，结算完成后若对所有目标角色均造成伤害，此【杀】不计入次数限制。",
 
 	guozhan_jun: "君主武将",
 	guozhan_jin: "文德武备",
