+++ mode/guozhan/src/translate/skill/character/index.js
@@ -3,6 +3,7 @@ import zhen from "./zhen.js";
 import shi from "./shi.js";
 import bian from "./bian.js";
 import quan from "./quan.js";
+import yingbian from "./yingbian.js";
 
 export default {
 	...normal,
@@ -10,4 +11,5 @@ export default {
 	...shi,
 	...bian,
 	...quan,
+	...yingbian,
 };
